pyproject.toml
[project]
name = "image-handle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.4",
    "pathlib>=1.0.1",
    "pillow>=12.0.0",
]
